完成作发布批改

+++ src/main/java/org/gdpi/course/service/ExamPaperService.java
@@ -65,5 +65,5 @@ public interface ExamPaperService {
      * @param grade
      * @param pid
      */
-    void addEssayGrade(Map<Integer, Integer> grade, Integer pid);
+    void addEssayGrade(Map<Integer, Float> grade, Integer pid);
 }
